% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addvecs}
\alias{addvecs}
\title{Adding Vectors}
\usage{
addvecs(...)
}
\arguments{
\item{...}{a list of vectors. All inputs must be of equal length.}
}
\value{
addvecs returns a vector of equal length to the inputs.
}
\description{
addvecs is used to add multiple vectors elementwise.
}
\examples{
temp <- seq(0,10)
x1 <- sample(temp, 3)
x2 <- sample(temp, 3)
addvecs(x1,x2)
addvecs(x1,x2,x1)

}
