% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dotprod}
\alias{dotprod}
\title{Dot Product}
\usage{
dotprod(...)
}
\arguments{
\item{input}{two vectors of equal length.}
}
\value{
dotprod returns the dot product of two vectors. That is \sum\limits_{i=1}^n x_i y_i.
}
\description{
Dot Product
}
\examples{
temp <- seq(0,10)
x1 <- sample(temp, 3)
x2 <- sample(temp, 3)
dotprod(x1,x2)
}
