% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multvecs}
\alias{multvecs}
\title{Multiplying Vectors Elementwise}
\usage{
multvecs(...)
}
\arguments{
\item{...}{a list of vectors. All inputs must be of equal length.}
}
\value{
multvecs returns a vector of equal length to the inputs.
}
\description{
multvecs is used to multiply a list of vectors elementwise.
}
\examples{
temp <- seq(0,10)
x1 <- sample(temp, 3)
x2 <- sample(temp, 3)
multvecs(x1,x2)
multvecs(x1,x2,x1)

}
